+++ tests/data/test_cc_oop/cc_products_active.html
@@ -376,12 +376,12 @@
    <p>
       The CCDB has approved a resolution to limit the validity of mutually recognized CC certificates over time.&nbsp; 
       Certificates will remain on the CPL for five years.&nbsp; 
-      Effective 1 June 2019, certificates with an expired validity period (that is, 5 years or more from the date of certificate issuance) will be moved to an Archive list on the CCRA portal
+      Effective 1 June 2019, certificates with an expired validity period (that is, 5 years or more from the date of sample issuance) will be moved to an Archive list on the CCRA portal
    </p>
    <p>
       The CCDB has approved a resolution to limit the validity of mutually recognized CC certificates over time.&nbsp; 
       Certificates will remain on the CPL for five years.&nbsp; 
-      Effective 1 June 2019, certificates with an expired validity period (that is, 5 years or more from the date of certificate issuance) will be moved to an Archive list on the CCRA portal.&nbsp; 
+      Effective 1 June 2019, certificates with an expired validity period (that is, 5 years or more from the date of sample issuance) will be moved to an Archive list on the CCRA portal.&nbsp;
       Certificates fully recognized under the CCRA must claim conformance to either a collaborative Protection Profile (cPP) or EAL1-2.&nbsp; 
    <p>
    -->
